@@ -1,76 +1,76 @@
 {
-  "name": "ionic-native",
-  "version": "4.7.0",
-  "description": "Native plugin wrappers for Cordova and Ionic with TypeScript, ES6+, Promise and Observable support",
-  "homepage": "https://ionicframework.com/",
-  "author": "Ionic Team <hi@ionic.io> (https://ionic.io)",
-  "license": "MIT",
-  "devDependencies": {
-    "@angular/compiler": "4.4.4",
-    "@angular/compiler-cli": "4.4.4",
-    "@angular/core": "4.4.4",
-    "@types/cordova": "0.0.34",
-    "@types/jasmine": "^2.6.3",
-    "@types/node": "^8.0.50",
-    "canonical-path": "0.0.2",
-    "child-process-promise": "2.2.1",
-    "conventional-changelog-cli": "1.3.4",
-    "cpr": "2.0.0",
-    "cz-conventional-changelog": "2.0.0",
-    "decamelize": "1.2.0",
-    "dgeni": "0.4.7",
-    "dgeni-packages": "0.16.10",
-    "fs-extra": "2.0.0",
-    "fs-extra-promise": "0.4.1",
-    "gulp": "3.9.1",
-    "gulp-rename": "1.2.2",
-    "gulp-replace": "0.5.4",
-    "gulp-tslint": "6.1.2",
-    "jasmine-core": "^2.6.1",
-    "karma": "^1.7.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-typescript": "^3.0.1",
-    "karma-typescript-es6-transform": "^1.0.0",
-    "lodash": "4.17.4",
-    "minimist": "1.2.0",
-    "node-html-encoder": "0.0.2",
-    "q": "1.5.0",
-    "queue": "4.2.1",
-    "rimraf": "2.6.1",
-    "rxjs": "^5.5.10",
-    "semver": "5.3.0",
-    "tslint": "3.15.1",
-    "tslint-ionic-rules": "0.0.8",
-    "typescript": "~2.4.2",
-    "zone.js": "^0.8.26"
-  },
-  "scripts": {
-    "start": "npm run test:watch",
-    "lint": "gulp lint",
-    "build": "npm run clean && npm run lint && npm run build:core && npm run build:modules",
-    "build:core": "ngc -p scripts/build/tsconfig-core.json",
-    "build:modules": "node scripts/build/build.js",
-    "clean": "rimraf dist .tmp",
-    "shipit": "npm run build && gulp readmes && npm run npmpub",
-    "npmpub": "node scripts/build/publish.js",
-    "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
-    "postchangelog": "git commit -am \"chore(): update changelog\"",
-    "test": "karma start karma.conf.js --single-run",
-    "test:watch": "karma start karma.conf.js",
-    "readmes": "gulp readmes"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ionic-team/ionic-native.git"
-  },
-  "bugs": {
-    "url": "https://github.com/ionic-team/ionic-native/issues"
-  },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
+    "name": "ionic-native",
+    "version": "4.7.0",
+    "description": "Native plugin wrappers for Cordova and Ionic with TypeScript, ES6+, Promise and Observable support",
+    "homepage": "https://ionicframework.com/",
+    "author": "Ionic Team <hi@ionic.io> (https://ionic.io)",
+    "license": "MIT",
+    "devDependencies": {
+        "@angular/compiler": "^6.0.0",
+        "@angular/compiler-cli": "^6.0.0",
+        "@angular/core": "^6.0.0",
+        "@types/cordova": "0.0.34",
+        "@types/jasmine": "^2.6.3",
+        "@types/node": "^8.0.50",
+        "canonical-path": "0.0.2",
+        "child-process-promise": "2.2.1",
+        "conventional-changelog-cli": "1.3.4",
+        "cpr": "2.0.0",
+        "cz-conventional-changelog": "2.0.0",
+        "decamelize": "1.2.0",
+        "dgeni": "0.4.7",
+        "dgeni-packages": "0.16.10",
+        "fs-extra": "2.0.0",
+        "fs-extra-promise": "0.4.1",
+        "gulp": "3.9.1",
+        "gulp-rename": "1.2.2",
+        "gulp-replace": "0.5.4",
+        "gulp-tslint": "6.1.2",
+        "jasmine-core": "^2.6.1",
+        "karma": "^1.7.0",
+        "karma-cli": "^1.0.1",
+        "karma-jasmine": "^1.1.0",
+        "karma-phantomjs-launcher": "^1.0.4",
+        "karma-typescript": "^3.0.1",
+        "karma-typescript-es6-transform": "^1.0.0",
+        "lodash": "4.17.4",
+        "minimist": "1.2.0",
+        "node-html-encoder": "0.0.2",
+        "q": "1.5.0",
+        "queue": "4.2.1",
+        "rimraf": "2.6.1",
+        "rxjs": "^6.2.1",
+        "semver": "5.3.0",
+        "tslint": "3.15.1",
+        "tslint-ionic-rules": "0.0.8",
+        "typescript": "~2.7.2",
+        "zone.js": "^0.8.26"
+    },
+    "scripts": {
+        "start": "npm run test:watch",
+        "lint": "gulp lint",
+        "build": "npm run clean && npm run lint && npm run build:core && npm run build:modules",
+        "build:core": "ngc -p scripts/build/tsconfig-core.json",
+        "build:modules": "node scripts/build/build.js",
+        "clean": "rimraf dist .tmp",
+        "shipit": "npm run build && gulp readmes && npm run npmpub",
+        "npmpub": "node scripts/build/publish.js",
+        "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
+        "postchangelog": "git commit -am \"chore(): update changelog\"",
+        "test": "karma start karma.conf.js --single-run",
+        "test:watch": "karma start karma.conf.js",
+        "readmes": "gulp readmes"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/ionic-team/ionic-native.git"
+    },
+    "bugs": {
+        "url": "https://github.com/ionic-team/ionic-native/issues"
+    },
+    "config": {
+        "commitizen": {
+            "path": "./node_modules/cz-conventional-changelog"
+        }
     }
-  }
 }
