@@ -1,6 +1,6 @@
 import { Plugin, Cordova, IonicNativePlugin, CordovaOptions, wrap } from '@ionic-native/core';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
+import { Observable } from 'rxjs';
+import { Observer } from 'rxjs';
 import { Injectable } from '@angular/core';
 
 /**
@@ -22,11 +22,11 @@ export interface CordovaFiniteObservableOptions extends CordovaOptions {
  *
  * Wraps method that returns an observable that can be completed. Provided opts.resultFinalPredicate dictates when the observable completes.
  */
-export function CordovaFiniteObservable(opts: CordovaFiniteObservableOptions = {}) {
+export function CordovaFiniteObservable (opts: CordovaFiniteObservableOptions = {}) {
   opts.observable = true;
   return (target: Object, methodName: string, descriptor: TypedPropertyDescriptor<any>) => {
     return {
-      value: function(...args: any[]) {
+      value: function (...args: any[]) {
         const wrappedObservable: Observable<any> = wrap(this, methodName, opts).apply(this, args);
         return new Observable<any>((observer: Observer<any>) => {
           const wrappedSubscription = wrappedObservable.subscribe({
@@ -108,7 +108,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     resultFinalPredicate: 'isLastChunk',
     resultTransform: 'library'
   })
-  getLibrary(options?: GetLibraryOptions): Observable<LibraryItem[]> { return; }
+  getLibrary (options?: GetLibraryOptions): Observable<LibraryItem[]> { return; }
 
   /**
    * Asks user permission to access photo library.
@@ -118,7 +118,7 @@ export class PhotoLibrary extends IonicNativePlugin {
   @Cordova({
     callbackOrder: 'reverse'
   })
-  requestAuthorization(options?: RequestAuthorizationOptions): Promise<void> { return; }
+  requestAuthorization (options?: RequestAuthorizationOptions): Promise<void> { return; }
 
   /**
    * Returns list of photo albums on device.
@@ -127,7 +127,7 @@ export class PhotoLibrary extends IonicNativePlugin {
   @Cordova({
     callbackOrder: 'reverse'
   })
-  getAlbums(): Promise<AlbumItem[]> { return; }
+  getAlbums (): Promise<AlbumItem[]> { return; }
 
   /**
    * Provides means to request URL of thumbnail, with specified size or quality.
@@ -139,7 +139,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     successIndex: 1,
     errorIndex: 2
   })
-  getThumbnailURL(photo: string | LibraryItem, options?: GetThumbnailOptions): Promise<string> { return; }
+  getThumbnailURL (photo: string | LibraryItem, options?: GetThumbnailOptions): Promise<string> { return; }
 
   /**
    * Provides means to request photo URL by id.
@@ -151,7 +151,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     successIndex: 1,
     errorIndex: 2
   })
-  getPhotoURL(photo: string | LibraryItem, options?: any): Promise<string> { return; }
+  getPhotoURL (photo: string | LibraryItem, options?: any): Promise<string> { return; }
 
   /**
    * Returns thumbnail as Blob.
@@ -163,7 +163,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     successIndex: 1,
     errorIndex: 2
   })
-  getThumbnail(photo: string | LibraryItem, options?: GetThumbnailOptions): Promise<Blob> { return; }
+  getThumbnail (photo: string | LibraryItem, options?: GetThumbnailOptions): Promise<Blob> { return; }
 
   /**
    * Returns photo as Blob.
@@ -175,7 +175,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     successIndex: 1,
     errorIndex: 2
   })
-  getPhoto(photo: string | LibraryItem, options?: any): Promise<Blob> { return; }
+  getPhoto (photo: string | LibraryItem, options?: any): Promise<Blob> { return; }
 
   /**
    * Saves image to specified album. Album will be created if not exists.
@@ -189,7 +189,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     successIndex: 2,
     errorIndex: 3
   })
-  saveImage(url: string, album: AlbumItem | string, options?: GetThumbnailOptions): Promise<LibraryItem> { return; }
+  saveImage (url: string, album: AlbumItem | string, options?: GetThumbnailOptions): Promise<LibraryItem> { return; }
 
   /**
    * Saves video to specified album. Album will be created if not exists.
@@ -201,7 +201,7 @@ export class PhotoLibrary extends IonicNativePlugin {
     successIndex: 2,
     errorIndex: 3
   })
-  saveVideo(url: string, album: AlbumItem | string): Promise<void> { return; }
+  saveVideo (url: string, album: AlbumItem | string): Promise<void> { return; }
 
 }
 
